requirements.txt
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
ecdsa==0.19.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
marshmallow==3.25.1
packaging==24.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
